+++ .eslintrc
@@ -13,6 +13,7 @@
     "browser": true,
     "shared-node-browser": true,
     "node": true,
+    "mocha": true,
     "es6": true
   },
   "settings": {
