+++ .mocharc.cjs
@@ -1,0 +1,5 @@
+module.exports = {
+  extension: ['js', 'jsx', 'ts', 'tsx'],
+  spec: ['tests/**.{js,ts,jsx,tsx}'],
+  loader: 'ts-node/esm'
+}
