+++ package.json
@@ -18,7 +18,7 @@
     "format": "prettier --check \"./**/*.ts\"",
     "format:fix": "prettier --write \"./**/*.ts\"",
     "lint": "eslint 'src/**/*.{js,jsx,ts,tsx}'",
-    "test": "react-scripts test",
+    "test": "mocha",
     "test:coverage": "npm run test -- --coverage --watchAll=false",
     "prepare": "husky install"
   },
@@ -48,6 +48,8 @@
     "@testing-library/react": "^12.0.0",
     "@testing-library/react-hooks": "^7.0.1",
     "@testing-library/user-event": "^13.2.1",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^16.4.13",
     "@typescript-eslint/eslint-plugin": "^4.28.5",
     "@typescript-eslint/parser": "^4.28.5",
     "bnc-onboard": "^1.32.0",
@@ -61,13 +63,14 @@
     "ethers": "^5.4.3",
     "ganache-cli": "^6.12.2",
     "husky": "^7.0.1",
+    "mocha": "^9.0.3",
     "prettier": "^2.3.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-scripts": "^4.0.3",
     "rollup": "^2.55.1",
     "rollup-plugin-filesize": "^9.1.1",
     "rollup-plugin-terser": "^7.0.2",
+    "ts-node": "^10.1.0",
     "tslib": "^2.3.0",
     "typescript": "^4.3.5"
   },
